+++ test/state.test.js
@@ -4,15 +4,27 @@ import type { Init
             , State
             , StateUpdate
             , Subscriptions } from "../src/state";
+import type { Message } from "../src/message";
 
 import ninos            from "ninos";
 import test             from "ava";
-import { updateData
+import { NONE
+       , updateData
        , updateAndSend } from "../src/update";
 import { subscribe } from "../src/message";
 
+// Type tests
+type MyMessage = { tag: "a" } | { tag: "b" };
+({
+  name: "test",
+  init: () => updateData("init"),
+  update: (data, msg) => NONE,
+  subscriptions: () => [],
+}: State<string, void, MyMessage>);
+
+
 test("State can be instantiated", t => {
-  const definition: State<string, void> = {
+  const definition: State<string, void, Message> = {
     name: "test",
     init: () => updateData("init"),
     update: (data, msg) => updateData(msg.tag),